backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
django-filter>=24.2
drf-spectacular>=0.27.2
gunicorn>=22.0.0
whitenoise>=6.6.0
dj-database-url>=2.2.0
django-cors-headers>=4.4.0
